Reorganize tests in classes
